Add files via upload
At present I am looking at finding information for one user, and have succeeded in getting the date and time of commits. This is complicated by the fact that commits depends on the user, the author and the specific repository name. This means there are many different variables needed to find commits, which could complicate the use of iteration. I think my next step will be to look at how I can find this information and efficiently store it.